Separates out point drawing.
