+++ Runtime/Scripts/ShapeDrawer.Point.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 98eebb56c8f2208499d4dfb1feaae702
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
